Initial commit. Pushing up a working password generator.
Changes still need to be done but a working password generator has been
created. I am really unsure of a proper folder structure for C# projects
but organized the files in a src folder and use the Program.cs in a
manner similar to an index.js or main.py file.

Since C# strictly adheres to OOP, the multiple files were used to
enhance readability and to make the code easier to maintain. This all
could have been written as a single class but I tried to split the best
of OOP and functional programming by creating multiple classes with
methods that have one purpose and chained these methods together later
to produce the functionality needed rather than writing long methods
in one class.