Adjusted the generation process to enforce additional checks to ensure
A secure password.

Min password length is now 16 characters.
Users can no longer select their character options as removing any
of the character sets will cause a weaker password to be generated.

The Main method has also been adjusted so that the program can be
run with arguments. The only supported argument is the required length
of the password to be generated. Invalid or no arguments will result in
the program prompting the user to enter a length, which will be
validated.